set redis flag

+++ controllers/landmarkController.js
@@ -2,7 +2,10 @@ const Landmark = require('./../models/landmarkModel');
 const catchAsync = require('./../utils/catchAsync');
 const AppError = require('./../utils/appError');
 const factory = require('./handlerFactory');
-const redisClient = require('./../utils/redisUtil');
+let redisClient;
+if (process.env.redis === 'true') {
+    redisClient = require('./../utils/redisUtil');
+}
 const multer = require('multer');
 const cloudinary = require('../utils/cloudinary');
 const streamifier = require('streamifier');
@@ -129,7 +132,9 @@ exports.updateLandmarkImages = catchAsync(async (req, res, next) => {
         landmark.imagesId[parseInt(imagesIndex[i])] = imagesId[i];
     }
     await landmark.save();
-    redisClient.flushAll();
+    if (process.env.redis === 'true') {
+        redisClient.flushAll();
+    }
     res.status(200).json({
         status: 'success',
         doc: {
@@ -174,7 +179,9 @@ exports.deleteLandmarkImages = catchAsync(async (req, res, next) => {
         landmark.imagesId = [];
     }
     await landmark.save();
-    redisClient.flushAll();
+    if (process.env.redis === 'true') {
+        redisClient.flushAll();
+    }
     res.status(200).json({
         status: 'success',
         doc: {
